Restore config
